% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{mapStyle}
\alias{mapStyle}
\title{Data on various Mapbox map styles.}
\format{
A tibble with 13 rows and 2 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{style} \tab Map style name \cr
   \code{url} \tab Map style URL \cr
}
}
\usage{
mapStyle
}
\description{
Data on various Mapbox map styles.
}
\examples{
mapStyle

}
\keyword{datasets}
