% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{student_country}
\alias{student_country}
\title{List of countries IHTM students come from by school year}
\format{
A tibble with 224 rows and 5 rows\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{cohort} \tab Cohort/class year \cr
   \code{country} \tab Name of country \cr
   \code{iso3c} \tab Country code \cr
   \code{lat} \tab Latitude coordinate \cr
   \code{lon} \tab Longitude coordinate \cr
}
}
\usage{
student_country
}
\description{
List of countries IHTM students come from by school year
}
\examples{
student_country

}
\keyword{datasets}
