% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{studentData}
\alias{studentData}
\title{Data on number of IHTM MSc students per country.}
\format{
A data frame with 28 rows and 4 columns:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{country} \tab Name of country location \cr
   \code{n} \tab Number of students \cr
   \code{lon} \tab Longitude coordinate \cr
   \code{lat} \tab Latitude coordinate \cr
}
}
\usage{
studentData
}
\description{
Data on number of IHTM MSc students per country.
}
\examples{
studentData

}
\keyword{datasets}
