% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{mapStyle}
\alias{mapStyle}
\title{mapStyle}
\format{A data frame with 2 variables
\describe{
\item{\code{style}}{Map style name}
\item{\code{url}}{Map style URL}
}}
\usage{
mapStyle
}
\description{
Data on various Mapbox map styles.
}
\keyword{datasets}
