% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{studentData}
\alias{studentData}
\title{studentData}
\format{A data frame with two variables:
\describe{
\item{\code{country}}{Name of country location}
\item{\code{n}}{Number of students}
\item{\code{lon}}{Longitude}
\item{\code{lat}}{Latitude}
}}
\usage{
studentData
}
\description{
Data on number of IHTM MSc students per country.
}
\keyword{datasets}
