% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infohub.R
\docType{data}
\name{studentPlacement}
\alias{studentPlacement}
\title{studentPlacement}
\format{A data frame with 7 variables:
\describe{
\item{\code{location}}{City/Area location of placement option}
\item{\code{country}}{Country location of placement option}
\item{\code{org_full}}{Name of organisation providing placement option (full)}
\item{\code{org_short}}{Name of organisation providing placement option (short)}
\item{\code{subject}}{Subject/topic for placement option}
\item{\code{keywords}}{Keywords/tags identifier for placement option}
\item{\code{url}}{Website/URL providing further information about placement option}
\item{\code{lon}}{Longitude}
\item{\code{lat}}{Latitude}
}}
\usage{
studentPlacement
}
\description{
Data on various placement options for IHTM MSc students.
}
\keyword{datasets}
